requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.31.0
python-dateutil==2.9.0
gunicorn==21.2.0
